Handle non-ASCII characters in log messages in Python 2
It seems that `dynamic_reconfigure` can handle non-ASCII characters just
fine, but the rospy loggers don't take unicode characters very well when
running in python 2. This change re-encodes to the native encoding and
replaces troublesome characters with '?' in the log messages.

On Python 3, the unicode characters are printed to the log just fine.