Handle invalid enum values
It seems to be possible for enum values to appear which aren't defined
in the configuration. I was able to set such a value using dynparam.

If the EnumEditor encounters one such value, add a temporary "invalid"
entry to the selector. If the current value is already invalid and we
get another one, just update that entry. Whenever a known value is sent
as an update, remove any temporary entries that happen to be present.

If everything works as it should, the user should never see an invalid
entry in the combo box unless it is the current value, and therefore
should never be able to change the enum to an invalid value.