Save instance state in rqt settings (#79)
New settings being saved:
- Current node list filter
- Currently expanded items in the node list
- Currently active editors (order preserved)

To reconcile the existing behavior for opening nodes from the command
line with the restoration of previously open nodes, I opted to treat the
nodes on the command line as an explicit list rather than an additive
one. So the command line list effectively 'overrides' the previous
state.

Signed-off-by: Scott K Logan <logans@cottsay.net>