Fix node selection from command line (#87)
This was sort-of working, but the tree of node names on the left side
got messed up when you tried to deselect a node. It was completely
ineffective for nodes in namespaces.

Turns out we were getting lucky that it worked at all, because the code
relied on storing a Qt index object that wasn't supposed to be
persistent. So when the tree got more complex, it completely fell apart.

I updated the code to just iterate over all of the items in the tree to
find the node we want. We'll only have to do this once, and only if the
user is specifying what nodes to open from the command line.

Forward-port of 9d256c40603ea8a9508d44bc1b2d797ac354bd64

Signed-off-by: Scott K Logan <logans@cottsay.net>