Use timeouts in service calls to avoid hangs (#98)
* Use async service calls to avoid hangs

Don't assume that all ROS nodes support the parameter services. Instead
explicitly look for the ListParameters service.

Also, don't assume that all nodes are spinning. Instead of using synchronous
service calls (which will hang if the node is not spinning to process its
messages) using an async call with a timeout.

Fixes #97, fixes #96, fixes #64

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>

* Use leading underscore on private function

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>

* Remove the sleep in the constructor and handle auto-refresh in a different PR

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>

* Cancel the future in case of timeout

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>

* Use rclpy.spin_until_future complete

Simpler code than the previous version which was based roughly off
of the code for client.call().

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>

* Address some code review feedback

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>